(internal) new package 'texts' (SplitLikeShell only package)
